pyproject.toml
[tool.poetry]
name = "project-hades"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "project_hades"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
jinja2 = "^3.1.2"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"
django-autoslug = "^1.9.8"
pillow = "^9.4.0"
django-filter = "^22.1"
django-froala-editor = "^4.0.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
